+++ app/upgrades/v1/testnet/btcstaking_params.go
@@ -22,5 +22,6 @@ const BtcStakingParamStr = `
   "min_unbonding_time_blocks": 0,
   "unbonding_fee_sat": "1000",
   "min_commission_rate": "0.03",
-  "delegation_creation_base_gas_fee": 1000
+  "delegation_creation_base_gas_fee": 1000,
+  "allow_list_expiration_height": 0
 }`