+++ testutil/btcstaking-helper/keeper.go
@@ -95,7 +95,7 @@ func NewHelper(
 	err = fk.SetParams(ctx, ftypes.DefaultParams())
 	require.NoError(t, err)
 
-	ctx = ctx.WithHeaderInfo(header.Info{Height: 1})
+	ctx = ctx.WithHeaderInfo(header.Info{Height: 1}).WithBlockHeight(1)
 
 	return &Helper{
 		t:   t,
@@ -137,7 +137,7 @@ func (h *Helper) BeginBlocker() {
 }
 
 func (h *Helper) GenAndApplyParams(r *rand.Rand) ([]*btcec.PrivateKey, []*btcec.PublicKey) {
-	return h.GenAndApplyCustomParams(r, 100, 0)
+	return h.GenAndApplyCustomParams(r, 100, 0, 0)
 }
 
 func (h *Helper) SetCtxHeight(height uint64) {
@@ -148,6 +148,7 @@ func (h *Helper) GenAndApplyCustomParams(
 	r *rand.Rand,
 	finalizationTimeout uint32,
 	minUnbondingTime uint32,
+	allowListExpirationHeight uint64,
 ) ([]*btcec.PrivateKey, []*btcec.PublicKey) {
 	// mock base header
 	baseHeader := btclctypes.SimnetGenesisBlock()
@@ -166,18 +167,19 @@ func (h *Helper) GenAndApplyCustomParams(
 	slashingPkScript, err := txscript.PayToAddrScript(slashingAddress)
 	h.NoError(err)
 	err = h.BTCStakingKeeper.SetParams(h.Ctx, types.Params{
-		CovenantPks:            bbn.NewBIP340PKsFromBTCPKs(covenantPKs),
-		CovenantQuorum:         3,
-		MinStakingValueSat:     1000,
-		MaxStakingValueSat:     int64(4 * 10e8),
-		MinStakingTimeBlocks:   10,
-		MaxStakingTimeBlocks:   10000,
-		SlashingPkScript:       slashingPkScript,
-		MinSlashingTxFeeSat:    10,
-		MinCommissionRate:      sdkmath.LegacyMustNewDecFromStr("0.01"),
-		SlashingRate:           sdkmath.LegacyNewDecWithPrec(int64(datagen.RandomInt(r, 41)+10), 2),
-		MinUnbondingTimeBlocks: minUnbondingTime,
-		UnbondingFeeSat:        1000,
+		CovenantPks:               bbn.NewBIP340PKsFromBTCPKs(covenantPKs),
+		CovenantQuorum:            3,
+		MinStakingValueSat:        1000,
+		MaxStakingValueSat:        int64(4 * 10e8),
+		MinStakingTimeBlocks:      10,
+		MaxStakingTimeBlocks:      10000,
+		SlashingPkScript:          slashingPkScript,
+		MinSlashingTxFeeSat:       10,
+		MinCommissionRate:         sdkmath.LegacyMustNewDecFromStr("0.01"),
+		SlashingRate:              sdkmath.LegacyNewDecWithPrec(int64(datagen.RandomInt(r, 41)+10), 2),
+		MinUnbondingTimeBlocks:    minUnbondingTime,
+		UnbondingFeeSat:           1000,
+		AllowListExpirationHeight: allowListExpirationHeight,
 	})
 	h.NoError(err)
 	return covenantSKs, covenantPKs
@@ -253,6 +255,7 @@ func (h *Helper) CreateDelegation(
 	unbondingValue int64,
 	unbondingTime uint16,
 	usePreApproval bool,
+	addToAllowList bool,
 ) (string, *types.MsgCreateBTCDelegation, *types.BTCDelegation, *btclctypes.BTCHeaderInfo, *types.InclusionProof, *UnbondingTxInfo, error) {
 	stakingTimeBlocks := stakingTime
 	bsParams := h.BTCStakingKeeper.GetParams(h.Ctx)
@@ -381,6 +384,10 @@ func (h *Helper) CreateDelegation(
 
 	if !usePreApproval {
 		msgCreateBTCDel.StakingTxInclusionProof = txInclusionProof
+	}
+
+	if addToAllowList {
+		h.BTCStakingKeeper.IndexAllowedStakingTransaction(h.Ctx, &stkTxHash)
 	}
 
 	_, err = h.MsgServer.CreateBTCDelegation(h.Ctx, msgCreateBTCDel)